requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
babel==2.17.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
#contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
docxcompose==1.4.0
docxtpl==0.19.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi>=0.100.0
fonttools==4.56.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
language_tool_python==2.9.3
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.0
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
pydantic>=2.7.0
pydantic_core>=2.0.0
Pygments==2.19.1
pyparsing==3.2.3
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart>=0.0.5
pytz==2024.2
#pywin32==308
#pywinauto>=0.6.8
pyzmq==26.2.0
requests==2.32.3
seaborn==0.13.2
selenium>=4.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.15.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
yarl==1.18.0
celery>=5.2.0
pydantic-settings>=2.0.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
pynput
redis>=4.0.0
gunicorn>=21.0.0
